Seperate GetSelfPath implementation for Blaze and Bazel
In Blaze, we keep using "/proc/self/exe", but in Bazel we resolve this symlink
to the actual Bazel binary.

This is essentially a rollback of
https://github.com/bazelbuild/bazel/commit/d79ec039a42feea4e4c059eca29f64b40aa8583a
only for Bazel, because it solved an issue that rare in Bazel,
but it's preventing users from running Bazel inside a Linux docker container on Apple Silicon machines.

Fixes https://github.com/bazelbuild/bazel/issues/15076

Closes https://github.com/bazelbuild/bazel/pull/14391

RELNOTES: None
PiperOrigin-RevId: 441177762
Change-Id: I48d7283cf7f42f4220e2261f02809c8b5270ef70